feat(carousel): provided styled and based classes for scroll, snap and swipe sliders
