+++ src/base/index.js
@@ -1,0 +1,3 @@
+export { default as BaseCarousel } from './BaseCarousel';
+export { default as BaseScrollSlider } from './BaseScrollSlider';
+export { default as BaseSwipeSlider } from './BaseSwipeSlider';
