+++ src/components/index.js
@@ -1,8 +1,7 @@
 export { default as Arrow } from './Arrow';
-export { default as BaseArrow } from './BaseArrow';
 export { default as Bullets } from './Bullets';
 export { default as Carousel } from './Carousel';
+export { default as SnapSlider } from './SnapSlider';
 export { default as ScrollSlider } from './ScrollSlider';
 export { default as UnevenItemsScrollSlider } from './UnevenItemsScrollSlider';
-export { default as SnapSlider } from './SnapSlider';
 export { default as SwipeSlider } from './SwipeSlider';
