+++ src/containers/ScrollSliderContainer.js
@@ -1,11 +1,11 @@
+import { BaseScrollSlider } from '#base';
 import { CarouselContext } from '#context';
-import { ScrollSlider } from '#components';
 import React from 'react';
 
 const ScrollSliderContainer = (props) => (
     <CarouselContext.Consumer>
         {({ activeItem, isRTL, itemsToShow, setItemsLength, goTo }) => (
-            <ScrollSlider
+            <BaseScrollSlider
                 {...props}
                 isRTL={isRTL}
                 activeItem={activeItem}
