+++ src/containers/index.js
@@ -1,4 +1,4 @@
 export { default as Bullets } from './BulletsContainer';
-export { default as SwipeSlider } from './SwipeSliderContainer';
-export { default as ScrollSlider } from './ScrollSliderContainer';
+export { default as BaseSwipeSlider } from './SwipeSliderContainer';
+export { default as BaseScrollSlider } from './ScrollSliderContainer';
 export { default as UnevenItemsScrollSlider } from './UnevenItemsScrollSliderContainer';
