+++ stories/SnapSliderCarousel.stories.jsx
@@ -1,6 +1,5 @@
-import '../styles.css';
 import './Carousel.css';
-import { Arrow, BaseArrow, Carousel, SnapSlider } from '../src';
+import { Arrow, Carousel, SnapSlider } from '../src';
 import React from 'react';
 
 export default {
@@ -71,15 +70,13 @@ const Template = (args) => (
                     <p>5</p>
                 </a>
             </SnapSlider>
-            <BaseArrow flow="prev">Prev</BaseArrow>
-            <BaseArrow flow="next">Next</BaseArrow>
-            {/* <Arrow flow={'prev'}>
+            <Arrow flow={'prev'}>
                 {({ onClick }) => <button onClick={onClick}>Previous</button>}
             </Arrow>
 
             <Arrow flow={'next'}>
                 {({ onClick }) => <button onClick={onClick}>Next</button>}
-            </Arrow> */}
+            </Arrow>
         </Carousel>
     </div>
 );
